requirements.txt
Django==3.0.3
django-cors-headers==3.7.0
djangorestframework==3.12.2
django-rest-swagger==2.2.0